Homepage function updated

+++ frontend-nextjs/frontend/travelsite/src/app/components/header.tsx
@@ -1,41 +1,62 @@
 'use client';
-import React from "react";
+import React, { useEffect } from "react";
 import './header.css';
 import { useRouter } from "next/navigation";
 import axios from "axios";
 import { useState } from "react";
 import { useAuth } from "./Authcontext";
+import { toast } from "react-toastify";
 
 
 
 const Header: React.FC = () => {
+  const [authenticated, setAuthenticated] = useState(false);
   const router = useRouter();
   const [errorMessage, setErrorMessage] = useState('');
   const {clearTimer} = useAuth();
 
+  useEffect(() => {
+    const token = localStorage.getItem('accesstoken');
+    setAuthenticated(!!token);
+
+  }, []);
 
   const logout = async () => {
+    if (authenticated === false) {
+      router.push('/signin');
+      return;
+      
+
+    }
     const token = localStorage.getItem('refreshtoken');
     if (token ) {
       try {
-        await axios.post('http://127.0.0.1:8000/auth/signout/',
-          { token }, {
-          headers: {
-            'Content-Type': 'application/json',
-            'Authorization': `Bearer ${token}`,
-          },
-          withCredentials: true
-        });
+        await toast.promise(
+          axios.post('http://127.0.0.1:8000/auth/signout/', { token }, {
+            headers: {
+              'Content-Type': 'application/json',
+              'Authorization': `Bearer ${token}`,
+            },
+            withCredentials: true
+          }),
+          {
+            pending: 'Signing out...',
+            success: 'Signed out successfully',
+            error: 'Error signing out'
+          }
+        );
 
         clearTimer();
         localStorage.clear();
-        setErrorMessage('');
         router.push('/signin');
       } catch (error: any) {
-        setErrorMessage(error.message);
-        console.error(error);
+        if(axios.isAxiosError(error) && error.response) {
+          toast.error(error.response.data.error || 'An error occurred');
+      } else {
+        toast.error('An error occurred');
       }
     }
+  }
   };
 
   const handleLogoClick = () => {
@@ -54,7 +75,7 @@ const Header: React.FC = () => {
           <div className="div-p">
             <p onClick={handlePClick}>Arya Vart</p>
           </div>
-          <button className="logout-button" onClick={logout}>SIGNOUT</button>
+          <button className="logout-button" onClick={logout}>{authenticated ? 'Sign out' : 'Login'}</button>
         </div>
       </div>
     </header>
